Day 4 completed
This day's project is about a simple game that the user has to guess a number between a minimum and maximum range in a total attempts (all of those numbers can be edited easily in the code as they are constant variables). The correct number is chosen randomly by the program.