Day 12 complete
This day's project is about a desktop application using the tkinter package, simulating a invoice manager for a restaurant. In the window, it is possible to see a list for food, drink and desserts and the user can select how much products wants. Then, click on invoice button to see the total price, the invoice and save it.